Move functions remaining in num to num-traits
Fixes #102.
